Add DebugMessageCallbackHandler
